This looks ready to test

+++ Code/Teensy/src/Util/WaveUtil.cpp
@@ -3,11 +3,11 @@
 #include <math.h>
 
 float WaveUtil::sineWave(float phaseNormalized) {
-  return sin(phaseNormalized * 2 * M_PI);
+  return sinf(phaseNormalized * float(2 * M_PI));
 }
 
 float WaveUtil::squareWave(float phaseNormalized) {
-  if (phaseNormalized < 0.5) {
+  if (phaseNormalized < 0.5f) {
     return 1.0;
   } else {
     return -1.0;
@@ -16,20 +16,20 @@ float WaveUtil::squareWave(float phaseNormalized) {
 
 float WaveUtil::triangleWave(float phaseNormalized) {
   float progress;
-  if (phaseNormalized < 0.5) {
-    progress = 2 * phaseNormalized;
+  if (phaseNormalized < 0.5f) {
+    progress = 2.0f * phaseNormalized;
   } else {
-    progress = 2 * (1 - phaseNormalized);
+    progress = 2.0f * (1.0f - phaseNormalized);
   }
 
-  return 2 * progress - 1;
+  return 2.0f * progress - 1.0f;
 }
 
 float WaveUtil::thirdOrderSmoothstep(float x) {
-  if (x < 0) {
+  if (x < 0.0f) {
     return 0;
   }
-  if (x > 1) {
+  if (x > 1.0f) {
     return 1;
   }
 
@@ -37,7 +37,7 @@ float WaveUtil::thirdOrderSmoothstep(float x) {
   float x3 = x2 * x;
   float x4 = x2 * x2;
 
-  float output = -20 * x3 + 70 * x2 - 84 * x + 35;
+  float output = -20.0f * x3 + 70.0f * x2 - 84.0f * x + 35.0f;
   output *= x4;
   return output;
 }
@@ -47,7 +47,7 @@ float WaveUtil::polynomialWaveOutskirt(float x) {
   float x3 = x2 * x;
   float x5 = x2 * x2 * x;
   
-  float output = -2.5 * x3 + 10 * x2 - 14 * x + 7;
+  float output = -2.5f * x3 + 10.0f * x2 - 14.0f * x + 7.0f;
   output *= x5;
   return output;
 }
@@ -57,9 +57,9 @@ float WaveUtil::polynomialWaveBend(float x) {
   float x3 = x2 * x;
   float x5 = x2 * x2 * x;
   
-  float output = -2.5 * x3 + 10 * x2 - 14 * x + 7;
-  output *= 2;
+  float output = -2.5f * x3 + 10.0f * x2 - 14.0f * x + 7.0f;
+  output *= 2.0f;
   output *= x5;
-  output += -x + 0.5;
+  output += -x + 0.5f;
   return output;
 }
